Moved action types and reducers into separate files
