+++ autoscheduler/frontend/src/components/CourseSelectCard/CollapsedCourseCard/CollapsedCourseCard.tsx
@@ -5,7 +5,7 @@ import {
 import ExpandIcon from '@material-ui/icons/ExpandMore';
 import { useSelector } from 'react-redux';
 import * as styles from './CollapsedCourseCard.css';
-import { RootState } from '../../../redux/reducers';
+import { RootState } from '../../../redux/reducer';
 
 interface CollapsedCourseCardProps {
     onExpand: Function;
