+++ autoscheduler/frontend/src/components/CourseSelectCard/ExpandedCourseCard/BasicSelect/BasicSelect.tsx
@@ -3,7 +3,7 @@ import {
   Checkbox, List, ListItem, ListItemIcon, ListItemText, FormLabel,
 } from '@material-ui/core';
 import { useSelector, useDispatch } from 'react-redux';
-import { RootState } from '../../../../redux/reducers';
+import { RootState } from '../../../../redux/reducer';
 import { updateCourseCard } from '../../../../redux/actions';
 
 interface BasicSelectProps {
