+++ autoscheduler/frontend/src/components/Schedule/ScheduleCard/ScheduleCard.tsx
@@ -2,7 +2,7 @@ import * as React from 'react';
 import { useSelector } from 'react-redux';
 import * as styles from './ScheduleCard.css';
 import DragHandle from './DragHandle';
-import { RootState } from '../../../redux/reducers';
+import { RootState } from '../../../redux/reducer';
 import { AvailabilityArgs } from '../../../types/Availability';
 import { formatTime } from '../../../timeUtil';
