+++ autoscheduler/frontend/src/tests/App.test.tsx
@@ -5,7 +5,7 @@ import { Provider } from 'react-redux';
 import { createStore } from 'redux';
 
 import App from '../components/App/App';
-import autoSchedulerReducer from '../redux/reducers';
+import autoSchedulerReducer from '../redux/reducer';
 
 test('renders without errors', () => {
   const store = createStore(autoSchedulerReducer);
