+++ autoscheduler/frontend/src/tests/Schedule.test.tsx
@@ -9,7 +9,7 @@ import Meeting, { MeetingType } from '../types/Meeting';
 import Section from '../types/Section';
 import Instructor from '../types/Instructor';
 import Schedule from '../components/Schedule/Schedule';
-import autoSchedulerReducer from '../redux/reducers';
+import autoSchedulerReducer from '../redux/reducer';
 
 const testSection = new Section({
   id: 123456,
